Restructure code to server and client packages
