+++ pkg/http-tunel/client/proxy.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by an AGPL-style
 // license that can be found in the LICENSE file.
 
-package tunnel
+package client
 
 import (
 	"github.com/NodeFactoryIo/vedran/pkg/http-tunel/proto"
