+++ pkg/http-tunel/server/addrpool.go
@@ -1,4 +1,4 @@
-package tunnel
+package server
 
 import (
 	"fmt"
